Refactored lift_score, confusion_matrix and sorted_feature_importance into metrics
