+++ src/ml_utils/metrics/__init__.py
@@ -1,5 +1,7 @@
-from .metrics import lift_score
+from .metrics import lift_score, confusion_matrix, sorted_feature_importance
 
 __all__ = [
-    'lift_score'
+    'lift_score',
+    'confusion_matrix',
+    'sorted_feature_importance'
 ]
