+++ src/ml_utils/metrics/metrics.py
@@ -1,4 +1,5 @@
 import numpy as np
+from sklearn import metrics
 
 
 class MetricError(Exception):
@@ -6,7 +7,7 @@ class MetricError(Exception):
 
 
 def lift_score(y_target, y_predicted):
-    if not isinstance(y_target, np.ndarray) or isinstance(y_predicted, np.ndarray):
+    if not isinstance(y_target, np.ndarray) or not isinstance(y_predicted, np.ndarray):
         raise MetricError("Input must be a numpy NDArray")
 
     n = len(y_target)
@@ -17,3 +18,15 @@ def lift_score(y_target, y_predicted):
     percent_correct_positives = (all_prod == 1).all(axis=1).sum() / n
 
     return percent_correct_positives / (percent_positives_target * percent_positives_predicted)
+
+
+def confusion_matrix(y_true, y_pred, normalized=True):
+    cm = metrics.confusion_matrix(y_true, y_pred)
+    if normalized is True:
+        cm = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
+    return cm
+
+
+def sorted_feature_importance(labels, importance):
+    idx = np.argsort(np.abs(importance))
+    return labels[idx], importance[idx]
