+++ src/ml_utils/visualizations/helpers.py
@@ -1,5 +1,4 @@
 import numpy as np
-from sklearn.metrics import confusion_matrix
 
 
 def cum_gain_curve(y_true, y_proba, positive_label=1):
@@ -10,11 +9,6 @@ def cum_gain_curve(y_true, y_proba, positive_label=1):
     cum_gains = np.cumsum(y_true[idx]) / n_true
     percents = np.arange(1, n + 1) / n
     return percents, cum_gains
-
-
-def sorted_feature_importance(labels, importance):
-    idx = np.argsort(np.abs(importance))
-    return labels[idx], importance[idx]
 
 
 def generate_text_labels(ax, horizontal=False, padding=0.005):
@@ -35,8 +29,3 @@ def generate_text_labels(ax, horizontal=False, padding=0.005):
         yield x_value, y_value
 
 
-def create_confusion_matrix(y_true, y_pred, normalized=True):
-    cm = confusion_matrix(y_true, y_pred)
-    if normalized is True:
-        cm = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
-    return cm
